Move userServices to user package. Update to save users to disk and also add support for updating password
